+++ tests/units/Glpi/System/Requirement/DbTimezones.php
@@ -34,25 +34,93 @@ namespace tests\units\Glpi\System\Requirement;
 
 class DbTimezones extends \GLPITestCase {
 
-   public function testCheckWithAvailableTimezones() {
+   public function testCheckWithUnavailableMysqlDb() {
 
       $this->mockGenerator->orphanize('__construct');
       $db = new \mock\DB();
-      $this->calling($db)->areTimezonesAvailable = true;
 
-      $this->newTestedInstance($db);
-      $this->boolean($this->testedInstance->isValidated())->isEqualTo(true);
-      $this->array($this->testedInstance->getValidationMessages())
-         ->isEqualTo(['Timezones seems loaded in database']);
-   }
+      $that = $this;
 
-   public function testCheckWithUnavailableTimezones() {
-
-      $this->mockGenerator->orphanize('__construct');
-      $db = new \mock\DB();
-      $this->calling($db)->areTimezonesAvailable = false;
+      $this->calling($db)->request = function ($query) use ($that) {
+         $result = new \mock\DBmysqlIterator(null);
+         if ($query === "SHOW DATABASES LIKE 'mysql'") {
+            $that->calling($result)->count = 0;
+         }
+         return $result;
+      };
 
       $this->newTestedInstance($db);
       $this->boolean($this->testedInstance->isValidated())->isEqualTo(false);
+      $this->array($this->testedInstance->getValidationMessages())->isEqualTo(['Access to timezone database (mysql) is not allowed.']);
+   }
+
+   public function testCheckWithUnavailableTimezonenameTable() {
+
+      $this->mockGenerator->orphanize('__construct');
+      $db = new \mock\DB();
+
+      $that = $this;
+
+      $this->calling($db)->request = function ($query) use ($that) {
+         $result = new \mock\DBmysqlIterator(null);
+         if ($query === "SHOW DATABASES LIKE 'mysql'") {
+            $that->calling($result)->count = 1;
+         } else if ($query === "SHOW TABLES FROM `mysql` LIKE 'time_zone_name'") {
+            $that->calling($result)->count = 0;
+         }
+         return $result;
+      };
+
+      $this->newTestedInstance($db);
+      $this->boolean($this->testedInstance->isValidated())->isEqualTo(false);
+      $this->array($this->testedInstance->getValidationMessages())->isEqualTo(['Access to timezone table (mysql.time_zone_name) is not allowed.']);
+   }
+
+   public function testCheckWithTimezonenameEmptyTable() {
+
+      $this->mockGenerator->orphanize('__construct');
+      $db = new \mock\DB();
+
+      $that = $this;
+
+      $this->calling($db)->request = function ($query) use ($that) {
+         $result = new \mock\DBmysqlIterator(null);
+         if ($query === "SHOW DATABASES LIKE 'mysql'") {
+            $that->calling($result)->count = 1;
+         } else if ($query === "SHOW TABLES FROM `mysql` LIKE 'time_zone_name'") {
+            $that->calling($result)->count = 1;
+         } else {
+            $that->calling($result)->current = ['cpt' => 0];
+         }
+         return $result;
+      };
+
+      $this->newTestedInstance($db);
+      $this->boolean($this->testedInstance->isValidated())->isEqualTo(false);
+      $this->array($this->testedInstance->getValidationMessages())->isEqualTo(['Timezones seems not loaded, see https://glpi-install.readthedocs.io/en/latest/timezones.html.']);
+   }
+
+   public function testCheckWithAvailableData() {
+
+      $this->mockGenerator->orphanize('__construct');
+      $db = new \mock\DB();
+
+      $that = $this;
+
+      $this->calling($db)->request = function ($query) use ($that) {
+         $result = new \mock\DBmysqlIterator(null);
+         if ($query === "SHOW DATABASES LIKE 'mysql'") {
+            $that->calling($result)->count = 1;
+         } else if ($query === "SHOW TABLES FROM `mysql` LIKE 'time_zone_name'") {
+            $that->calling($result)->count = 1;
+         } else {
+            $that->calling($result)->current = ['cpt' => 30];
+         }
+         return $result;
+      };
+
+      $this->newTestedInstance($db);
+      $this->boolean($this->testedInstance->isValidated())->isEqualTo(true);
+      $this->array($this->testedInstance->getValidationMessages())->isEqualTo(['Timezones seems loaded in database.']);
    }
 }